+++ core/config/services/controller.yml
@@ -42,41 +42,8 @@ services:
             - @request
             - @dbal.conn
 
-    consim.core.controller.acpnews:
-        class: consim\core\controller\ACP_News
-        arguments:
-            - @config
-            - @service_container
-            - @controller.helper
-            - @user
-            - @template
-            - @request
-            - @dbal.conn
-
-    consim.core.controller.acpanniversary:
-        class: consim\core\controller\ACP_Anniversary
-        arguments:
-            - @config
-            - @service_container
-            - @controller.helper
-            - @user
-            - @template
-            - @request
-            - @dbal.conn
-
     consim.core.controller.map:
-            class: consim\core\controller\Map
-            arguments:
-                - @config
-                - @service_container
-                - @controller.helper
-                - @user
-                - @template
-                - @request
-                - @dbal.conn
-
-    consim.core.controller.acpmap:
-        class: consim\core\controller\ACP_Map
+        class: consim\core\controller\MapController
         arguments:
             - @config
             - @service_container
