+++ core/service/Inventories.php
@@ -6,7 +6,7 @@
  * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
  */
 
-namespace consim\core\operators;
+namespace consim\core\service;
 
 use consim\core\entity\InventoryItem;
 use Symfony\Component\DependencyInjection\ContainerInterface;
